List Valid Google Language Codes

Gives a list of valid Language Codes for Google Fonts, or validates that the language codes given are valid.

Usage

google_language_codes(
  language_codes = c("latin", "latin-ext", "sinhala", "greek", "hebrew", "vietnamese",
    "cyrillic", "cyrillic-ext", "devanagari", "arabic", "khmer", "tamil", "greek-ext",
    "thai", "bengali", "gujarati", "oriya", "malayalam", "gurmukhi", "kannada", "telugu",
    "myanmar")
)

Arguments

language_codes

Vector of potential Google language codes

Value

A vector of Google Font language codes matching language_codes.

See also