xaringanthemer: Xaringan CSS Theme Generator

Create and edit your xaringan CSS files from within your R R Markdown slide source. More technically, this package actually creates valid remarkjs CSS themes that are easily used by xarangan.

See also