Write A Customized Xaringan Theme

Creates a customized Xaringan theme CSS file.

style_xaringan(
  text_color = "#000",
  header_color = "#000",
  background_color = "#FFF",
  link_color = "rgb(249, 38, 114)",
  text_bold_color = NULL,
  text_slide_number_color = inverse_background_color,
  padding = "16px 64px 16px 64px",
  background_image = NULL,
  background_size = NULL,
  background_position = NULL,
  code_highlight_color = "rgba(255,255,0,0.5)",
  code_inline_color = "#000",
  code_inline_background_color = NULL,
  code_inline_font_size = "1em",
  inverse_background_color = "#272822",
  inverse_text_color = "#d6d6d6",
  inverse_text_shadow = FALSE,
  inverse_header_color = "#f3f3f3",
  title_slide_text_color = inverse_text_color,
  title_slide_background_color = inverse_background_color,
  title_slide_background_image = NULL,
  title_slide_background_size = NULL,
  title_slide_background_position = NULL,
  footnote_color = NULL,
  footnote_font_size = "0.9em",
  footnote_position_bottom = "60px",
  left_column_subtle_color = "#777",
  left_column_selected_color = "#000",
  blockquote_left_border_color = "lightgray",
  table_border_color = "#666",
  table_row_border_color = "#ddd",
  table_row_even_background_color = "#eee",
  base_font_size = "20px",
  text_font_size = "1rem",
  header_h1_font_size = "2.75rem",
  header_h2_font_size = "2.25rem",
  header_h3_font_size = "1.75rem",
  header_background_auto = FALSE,
  header_background_color = header_color,
  header_background_text_color = background_color,
  header_background_padding = NULL,
  header_background_content_padding_top = "7rem",
  header_background_ignore_classes = c("normal", "inverse", "title", "middle", "bottom"),
  text_slide_number_font_size = "0.9em",
  text_font_google = NULL,
  text_font_family = "Noto Sans",
  text_font_weight = "normal",

    text_font_url = "https://fonts.googleapis.com/css?family=Noto+Sans:400,400i,700,700i&display=swap",

    text_font_family_fallback = "-apple-system, BlinkMacSystemFont, avenir next, avenir, helvetica neue, helvetica, Ubuntu, roboto, noto, segoe ui, arial",
  text_font_base = "sans-serif",
  header_font_google = NULL,
  header_font_family = "Cabin",
  header_font_weight = "600",

    header_font_url = "https://fonts.googleapis.com/css?family=Cabin:600,600i&display=swap",
  code_font_google = NULL,
  code_font_family = "Source Code Pro",
  code_font_size = "0.9rem",

    code_font_url = "https://fonts.googleapis.com/css?family=Source+Code+Pro:400,700&display=swap",
  code_font_family_fallback = "Menlo, Consolas, Monaco, Liberation Mono, Lucida Console",
  colors = NULL,
  extra_css = NULL,
  extra_fonts = NULL,
  outfile = "xaringan-themer.css"
)

Arguments

text_color
Text Color. Defaults to #000. Modifies the body element.

header_color
Header Color. Defaults to #000. Modifies the h1, h2, h3 elements.

background_color
Slide Background Color. Defaults to #FFF. Modifies the .remark-slide-content class.

link_color
Link Color. Defaults to rgb(249, 38, 114). Modifies the a, a > code elements.

text_bold_color
Bold Text Color. Defaults to NULL. Modifies the strong element.

text_slide_number_color
Slide Number Color. Defaults to inverse_background_color. Modifies the .remark-slide-number class.

padding
Slide Padding in top right [bottom left] format. Defaults to 16px 64px 16px 64px. Modifies the .remark-slide-content class.

background_image
Background image applied to each and every slide. Set title_slide_background_image = "none" to remove the background image from the title slide. Defaults to NULL. Modifies the .remark-slide-content class.

background_size
Background image size, requires background_image to be set. If background_image is set, background_size will default to cover so the backround fills the screen. If both background_image and background_position are set, will default to 100 percent. Defaults to NULL. Modifies the .remark-slide-content class.

background_position
Background image position, requires background_image to be set, and it is recommended to adjust background_size. Defaults to NULL. Modifies the .remark-slide-content class.

code_highlight_color
Code Line Highlight. Defaults to rgba(255,255,0,0.5). Modifies the .remark-code-line-highlighted class.

code_inline_color
Inline Code Color. Defaults to #000. Modifies the .remark-inline-code class.

code_inline_background_color
Inline Code Background Color. Defaults to NULL. Modifies the .remark-inline-code class.

code_inline_font_size
Inline Code Text Font Size. Defaults to 1em. Modifies the .remark-inline-code class.

inverse_background_color
Inverse Background Color. Defaults to #272822. Modifies the .inverse class.

inverse_text_color
Inverse Text Color. Defaults to #d6d6d6. Modifies the .inverse class.

inverse_text_shadow
Enables Shadow on text of inverse slides. Defaults to FALSE. Modifies the .inverse class.

inverse_header_color
Inverse Header Color. Defaults to #f3f3f3. Modifies the .inverse h1, .inverse h2, .inverse h3 classes.

title_slide_text_color
Title Slide Text Color. Defaults to inverse_text_color. Modifies the .title-slide class.

title_slide_background_color
Title Slide Background Color. Defaults to inverse_background_color. Modifies the .title-slide class.

title_slide_background_image
Title Slide Background Image URL. Defaults to NULL. Modifies the .title-slide class.

title_slide_background_size
Title Slide Background Image Size, defaults to "cover" if background image is set. Defaults to NULL. Modifies the .title-slide class.

title_slide_background_position
Title Slide Background Image Position. Defaults to NULL. Modifies the .title-slide class.

footnote_color
Footnote text color (if NA, then it will be the same color as text_color). Defaults to NULL. Modifies the .footnote class.

footnote_font_size
Footnote font size. Defaults to 0.9em. Modifies the .footnote class.

footnote_position_bottom
Footnote location from bottom of screen. Defaults to 60px. Modifies the .footnote class.

left_column_subtle_color
Left Column Text (not last). Defaults to #777. Modifies the .left-column h2, .left-column h3 classes.

left_column_selected_color
Left Column Current Selection. Defaults to #000. Modifies the .left-column h2:last-of-type, .left-column h3:last-child classes.

blockquote_left_border_color
Blockquote Left Border Color. Defaults to lightgray. Modifies the blockquote element.

table_border_color
Table top/bottom border. Defaults to #666. Modifies the table: border-top, border-bottom elements.

table_row_border_color
Table row inner bottom border. Defaults to #ddd. Modifies the table thead th: border-bottom elements.

table_row_even_background_color
Table Even Row Background Color. Defaults to #eee. Modifies the thead, tfoot, tr:nth-child(even) elements.

base_font_size
Base Font Size for All Slide Elements (must be px). Defaults to 20px. Modifies the html element.

text_font_size
Slide Body Text Font Size. Defaults to 1rem. Modifies the .remark-slide-content class.

header_h1_font_size
h1 Header Text Font Size. Defaults to 2.75rem. Modifies the .remark-slide-content h1 class.

header_h2_font_size
h2 Header Text Font Size. Defaults to 2.25rem. Modifies the .remark-slide-content h2 class.

header_h3_font_size
h3 Header Text Font Size. Defaults to 1.75rem. Modifies the .remark-slide-content h3 class.

header_background_auto
Add background under slide title automatically for h1 header elements. If not enabled, use class: header_background to enable. Defaults to FALSE.

header_background_color
Background Color for h1 Header with Background. Defaults to header_color. Modifies the .remark-slide-content h1 class.

header_background_text_color
Text Color for h1 Header with Background. Defaults to background_color. Modifies the .remark-slide-content h1 class.

header_background_padding
Padding for h1 Header with Background. Defaults to NULL. Modifies the .remark-slide-content h1 class.

header_background_content_padding_top
Top Padding for Content in Slide with Header with Background. Defaults to 7rem. Modifies the .remark-slide-content class.

header_background_ignore_classes
Slide Classes Where Header with Background will not be Applied. Defaults to c('normal', 'inverse', 'title', 'middle', 'bottom'). Modifies the .remark-slide-content class.

text_slide_number_font_size
Slide Number Text Font Size. Defaults to 0.9em. Modifies the .remark-slide-number class.

text_font_google
Use google_font() to specify body font. Defaults to NULL. Modifies the body element.

text_font_family
Body Text Font Family (xaringan default is 'Droid Serif'). Defaults to Noto Sans. Modifies the body element.

text_font_weight
Body Text Font Weight. Defaults to normal. Modifies the body element.

text_font_url
Body Text Font URL(s). Defaults to https://fonts.googleapis.com/css?family=Noto+Sans:400,400i,700,700i&display=swap. Modifies the @import url() elements.

text_font_family_fallback
Body Text Font Fallbacks. Defaults to -apple-system, BlinkMacSystemFont, avenir next, avenir, helvetica neue, helvetica, Ubuntu, roboto, noto, segoe ui, arial. Modifies the body element.

text_font_base
Body Text Base Font (Total Failure Fallback). Defaults to sans-serif. Modifies the body element.

header_font_google
Use google_font() to specify header font. Defaults to NULL. Modifies the body element.

header_font_family
Header Font Family (xaringan default is 'Yanone Kaffeesatz'). Defaults to Cabin. Modifies the h1, h2, h3 elements.

header_font_weight
Header Font Weight. Defaults to 600. Modifies the h1, h2, h3 elements.

header_font_url
Header Font URL. Defaults to https://fonts.googleapis.com/css?family=Cabin:600,600i&display=swap. Modifies the @import url elements.

code_font_google
Use google_font() to specify code font. Defaults to NULL. Modifies the body element.

code_font_family
Code Font Family. Defaults to Source Code Pro. Modifies the .remark-code, .remark-inline-code classes.

code_font_size
Code Text Font Size. Defaults to 0.9rem. Modifies the .remark-inline class.

code_font_url
Code Font URL. Defaults to https://fonts.googleapis.com/css?family=Source+Code+Pro:400,700&display=swap. Modifies the @import url elements.

code_font_family_fallback
Code Font Fallback. Defaults to Menlo, Consolas, Monaco, Liberation Mono, Lucida Console. Modifies the .remark-code, .remark-inline-code classes.

colors
A named vector of custom colors. The names of the colors become CSS variables and classes that can be used within your slides. For example, colors = c(blue = "#bad4ed") adds a CSS variable --blue, a .blue CSS class that applies the color to the text or foreground color, and a .bg-blue CSS class that applies the color to the background.

extra_css
A named list of CSS definitions each containing a named list of CSS property-value pairs, i.e. list(".class-id" = list("css-property" = "value")).

extra_fonts
A list of additional fonts to import, each list element can be either a URL as a character string or a call to google_font(). To use a font imported in extra_fonts, you will need to write custom CSS rules that apply the font to an element or class with the font-family property. See the Fonts section of vignette("xaringanthemer") for an example.

outfile
Customized xaringan CSS output file name, default is "xaringan-themer.css"

See also